Improve backoffice quality guardrails
