+++ package.json
@@ -36,11 +36,12 @@
     "migrate:db": "bash scripts/migrate-db.sh",
     "populate-db": "npm run dev --workspace=services/movie-seed",
     "prepare": "husky || true",
+    "quality:backoffice": "npm run quality:structure --workspace=apps/backoffice",
     "setup:local-db": "bash scripts/setup-local-db.sh",
     "start:storybook": "npm run start:storybook --workspace=apps/web",
     "storybook": "npm run storybook --workspace=apps/web",
     "test": "npm run test --workspace=apps/web",
-    "test:backoffice": "npm run build --workspace=packages/shared && npm run test --workspace=apps/backoffice",
+    "test:backoffice": "npm run build --workspace=packages/shared && npm run quality:backoffice && npm run test --workspace=apps/backoffice",
     "test:e2e": "npm run test:e2e:setup && npm run test:e2e:run",
     "test:e2e:down": "node scripts/e2e/teardown-db.mjs",
     "test:e2e:run": "playwright test --config apps/web/playwright.e2e.config.ts",
